Connected and Merged NoteSpace component with Notes and NoteEditor - To share states between Note and NoteEditor, we need a parent component, thus NoteSpace is being made with that in mind. - Now when pressing the edit button an edit form should open prefilled with data from the Note component (See ActiveNote state). - Other small changes and fixes to frontend.
